Moved encrypted password method to service
