--- internal/service/user/utils.go
@@ -1,12 +1,0 @@
-package usersvc
-
-import "golang.org/x/crypto/bcrypt"
-
-func encryptedPassword(password string) (string, error) {
-	passwd, err := bcrypt.GenerateFromPassword([]byte(password), bcrypt.DefaultCost)
-	if err != nil {
-		return "", err
-	}
-
-	return string(passwd), nil
-}
